# Faster implementation of {Int,Int32,Int64,Nativeint}.to_string