--- a/stdlib/int.ml
+++ b/stdlib/int.ml
@@ -47,9 +47,10 @@ external of_float : float -> int = "%intoffloat"
 external int_of_string : string -> int = "caml_int_of_string"
 let of_string s = try Some (int_of_string s) with Failure _ -> None
 *)
-
+(*
 external format_int : string -> int -> string = "caml_format_int"
-let to_string x = format_int "%d" x
+*)
+external to_string : int -> string = "caml_int_to_string"
 
 external seeded_hash_param :
   int -> int -> int -> 'a -> int = "caml_hash" [@@noalloc]