@gasche 2023-04-15
The performance of `sprintf` depends on your libc. Which OS / libc were you using in your performance tests?