@atrieu 2023-04-15
I have tested on a macOS machine with `Apple clang version 14.0.0 (clang-1400.0.29.202)`.
Also tested on a Linux machine with `gcc 12.2.0` and `glibc 2.36`.